Async load scene files.
The problem: The client uses the main thread to send pings to the server.
On  weak hardware, when it takes a long time to load the scene, it doesn't send
pings to the server and the client may think it disconnected.

Now the scene is loaded in a worker thread and the main thread is able to ping
the server.